add env for port and content

+++ run_quartz.sh
@@ -7,9 +7,8 @@ sudo npm i npm@11.2.0 -g
 sudo apt update
 sudo apt install nginx
 
-sudo unlink /etc/nginx/sites-enabled/default
+if [[ -f /etc/nginx/sites-enabled/default ]] ; then sudo unlink /etc/nginx/sites-enabled/default ;fi
 sudo mkdir -p /etc/nginx/locations
-
 
 
 # https://stackoverflow.com/questions/59895/how-do-i-get-the-directory-where-a-bash-script-is-located-from-within-the-script
@@ -21,10 +20,15 @@ while [ -L "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symli
 done
 DIR=$( cd -P "$( dirname "$SOURCE" )" >/dev/null 2>&1 && pwd )
 
-
+# env 
+if [[ ! -z "$QUARTZ_PORT" ]]; then echo QUARTZ_PORT $QUARTZ_PORT; else QUARTZ_PORT=8000 ;fi
+if [[ ! -z "$QUARTZ_CONTENT" ]]; then echo QUARTZ_CONTENT $QUARTZ_CONTENT; else QUARTZ_CONTENT=$DIR/content ;fi
+echo QUARTZ_PORT $QUARTZ_PORT
+echo QUARTZ_CONTENT $QUARTZ_CONTENT
 
 sudo cp $DIR/default.conf /etc/nginx/conf.d/default.conf
 sudo cp $DIR/location-*.conf /etc/nginx/locations/
+sudo sed -i  "/proxy_pass/s/127.0.0.1:[0-9]\+/127.0.0.1:$QUARTZ_PORT/" /etc/nginx/locations/location-quartz.conf
 
 sudo service nginx restart
 sudo nginx -t && sudo systemctl reload nginx
@@ -32,10 +36,11 @@ sudo nginx -t && sudo systemctl reload nginx
 # run app
 
 npm i --prefix $DIR
-#npx quartz build --serve --port 8002
-cd $DIR &&  npx quartz build  --serve --watch --port 8002  -d $DIR/content
 
-echo "commit github update"
+cd $DIR &&  npx quartz build  --serve --watch --port $QUARTZ_PORT  -d $QUARTZ_CONTENT
+
+
+#echo "commit github update"
 
 #cd $DIR
 #git add --all
